backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-multipart
openai
python-dotenv